sandbox/requirements.txt
flask==3.0.0
selenium==4.15.0
gunicorn==21.2.0
pillow==10.1.0
